+++ Dockerfile
@@ -38,6 +38,7 @@ COPY docker/nginx.conf /etc/netbox-nginx/nginx.conf
 COPY docker/docker-entrypoint.sh docker-entrypoint.sh
 COPY startup_scripts/ /opt/netbox/startup_scripts/
 COPY initializers/ /opt/netbox/initializers/
+COPY configuration/configuration.py /etc/netbox/configuration.py
 
 WORKDIR /opt/netbox/netbox
