+++ docker/gunicorn_config.py
@@ -2,7 +2,6 @@ command = '/usr/bin/gunicorn'
 pythonpath = '/opt/netbox/netbox'
 bind = '0.0.0.0:8001'
 workers = 3
-user = 'root'
 errorlog = '-'
 accesslog = '-'
 capture_output = False
